restructured project and updated references
