+++ character-management/default-json-configs/barbarian.json
