+++ character-management/default-json-configs/bard.json
