+++ character-management/default-json-configs/character.json
