+++ character-management/default-json-configs/cleric.json
