+++ character-management/default-json-configs/druid.json
