+++ character-management/default-json-configs/embed.go
