+++ character-management/default-json-configs/ranger.json
