+++ character-management/default-json-configs/rogue.json
