+++ character-management/models/character.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"strings"
 
+	"github.com/onioncall/dndgo/character-management/types"
 	"github.com/onioncall/dndgo/logger"
-	"github.com/onioncall/dndgo/types"
 )
 
 type Character struct {
