+++ character-management/models/class/bard.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"math"
 
+	"github.com/onioncall/dndgo/character-management/models"
 	"github.com/onioncall/dndgo/logger"
-	"github.com/onioncall/dndgo/models"
 )
 
 type Bard struct {
