+++ character-management/models/class/cleric.go
@@ -4,9 +4,9 @@ import (
 	"encoding/json"
 	"fmt"
 
+	"github.com/onioncall/dndgo/character-management/models"
+	"github.com/onioncall/dndgo/character-management/types"
 	"github.com/onioncall/dndgo/logger"
-	"github.com/onioncall/dndgo/models"
-	"github.com/onioncall/dndgo/types"
 )
 
 type Cleric struct {
