+++ character-management/types/abilities.go
