+++ character-management/types/character.go
