+++ character-management/types/equipment.go
