+++ character-management/types/skills.go
