+++ character-management/types/spells.go
