+++ character-management/types/weapons.go
