+++ search/api/client.go
