+++ search/api/responses/equipment.go
