+++ search/api/responses/shared.go
