+++ search/api/responses/spells.go
